+++ src/app/elements/chat/chat.component.html
@@ -1,6 +1,8 @@
 <div id="chatContainer" class="chat-container">
   <nz-float-button-group [hidden]="chatAIShown">
-    <nz-float-button (click)="handleShowDrawer()" [nzIcon]="setting"></nz-float-button>
+    @if(isShowSetting && subViews.length === 1) {
+      <nz-float-button (click)="handleShowDrawer()" [nzIcon]="setting"></nz-float-button>
+    }
     <nz-float-button (click)="showChatAI()" [nzIcon]="customer"> </nz-float-button>
   </nz-float-button-group>
   <ng-template #setting>
