+++ src/app/elements/drawer/drawer.component.scss
@@ -1,5 +1,5 @@
 ::ng-deep .elements-drawer-wrapper {
-  background-color: white;
+  background-color: var(--el-drawer-bg-color);
   width: 600px;
   height: calc(100vh - 60px);
   overflow: auto;
@@ -14,6 +14,42 @@
     top: 20px;
     right: 20px;
     z-index: 1000;
+
+    svg {
+      color: var(--el-text-color-light);
+    }
+  }
+
+  .ant-tabs {
+    .ant-tabs-nav .ant-tabs-nav-list {
+      width: 100%;
+
+      .ant-tabs-tab {
+        width: 50%;
+        justify-content: center;
+        color: var(--el-text-color-light);
+      }
+    }
+
+    .ant-tabs-content-holder {
+      .ant-select {
+        .ant-select-selector {
+          background-color: var(--el-drawer-collapse-bg-color);
+          border: 1px solid var(--el-divider-border-color);
+          border-radius: 6px;
+          color: var(--el-text-color-light);
+
+          &:hover {
+            border-color: var(--el-form-hover-border-color);
+            background-color: var(--el-drawer-bg-color);
+          }
+        }
+
+        .ant-select-arrow {
+          color: var(--el-text-color-light);
+        }
+      }
+    }
   }
 
   .ant-card {
@@ -57,137 +93,5 @@
       }
     }
 
-    .ant-tabs {
-      .ant-tabs-nav {
-        .ant-tabs-nav-list {
-          width: 100%;
-
-          .ant-tabs-tab {
-            width: 50%;
-            justify-content: center;
-            color: var(--el-text-color-light);
-          }
-        }
-      }
-
-      .ant-tabs-content-holder {
-        .setting-form {
-          label {
-            color: var(--el-text-color-light);
-          }
-
-          .ant-select {
-            .ant-select-selector {
-              background-color: var(--el-drawer-collapse-bg-color);
-              border: 1px solid var(--el-divider-border-color);
-              border-radius: 6px;
-              color: var(--el-text-color-light);
-
-              &:hover {
-                border-color: var(--el-form-hover-border-color);
-                background-color: var(--el-drawer-bg-color);
-              }
-            }
-
-            .ant-select-arrow {
-              color: var(--el-text-color-light);
-            }
-          }
-        }
-
-        .share-button {
-          border-radius: 4px;
-        }
-
-        .share-user-list {
-          height: calc(100vh - 152px);
-        }
-
-        .file-manager-container {
-          width: 100%;
-          position: relative;
-
-          .file-manager-loading {
-            display: flex;
-            align-items: center;
-            justify-content: center;
-            height: 100%;
-            color: var(--el-text-color-light);
-          }
-        }
-
-        .ant-empty {
-          color: var(--el-text-color-light);
-        }
-
-        .ant-tag {
-          margin: 0;
-          font-size: 11px;
-          padding: 0 6px;
-          height: 20px;
-          line-height: 18px;
-          border-radius: 4px;
-        }
-
-        .ant-divider {
-          margin: 0 0 12px 0;
-          border-top-color: var(--el-divider-border-color);
-
-          ::ng-deep .ant-divider-inner-text {
-            display: flex;
-            align-items: center;
-            gap: 8px;
-            padding: 0 12px;
-            font-size: 14px;
-            font-weight: 500;
-            color: var(--el-text-color-light);
-          }
-        }
-
-        .ant-list {
-          .ant-list-item-meta {
-            align-items: center;
-
-            .ant-list-item-meta-content {
-              .ant-list-item-meta-description,
-              .ant-list-item-meta-title {
-                margin-bottom: unset;
-                color: var(--el-text-color-light);
-              }
-            }
-          }
-
-          .ant-list-item-action li {
-            color: var(--el-text-color-light);
-
-            svg {
-              transition: color 0.3s ease;
-
-              &:hover {
-                cursor: pointer;
-                color: red;
-              }
-            }
-          }
-        }
-      }
-    }
-  }
-}
-
-::ng-deep .create-share-link-modal {
-  .ant-modal-content {
-    border-radius: 4px;
-    overflow: hidden;
-
-    .ant-modal-header {
-      border-radius: 8px 8px 0 0;
-    }
-  }
-}
-
-::ng-deep .share-user-delete-button {
-  .ant-popover-inner {
-    border-radius: 4px;
   }
 }
