+++ src/app/elements/drawer/filemanager/filemanager.component.html
@@ -1,0 +1,6 @@
+<section>
+  <div
+    #iframeContainer
+    style="width: 100%; height: calc(100vh - 170px); position: relative;">
+  </div>
+</section>
