+++ src/app/elements/drawer/session-share/session-share.component.html
@@ -1,0 +1,175 @@
+<section>
+  <div class="share-user-list">
+    <nz-divider nzDashed [nzText]="text" nzOrientation="left">
+      <ng-template #text>
+        <nz-icon nzType="team" nzTheme="outline" />
+        {{ 'OnlineUser' | translate }} ({{ onLineUsers().length }})
+      </ng-template>
+    </nz-divider>
+
+    <div style="margin-bottom: 16px">
+      @for (item of onLineUsers(); track item.terminal_id + item.user + item.primary) {
+        <nz-card
+          nzSize="small"
+          style="margin-bottom: 8px; border-radius: 8px"
+          [nzBodyStyle]="{ padding: '12px 16px' }"
+        >
+          <div style="display: flex; align-items: center; justify-content: space-between">
+            <div>
+              <div style="font-weight: 500; margin-bottom: 4px; color: var(--el-text-color-light)">
+                {{ item.user }}
+              </div>
+              <div style="display: flex; gap: 4px">
+                <nz-tag *ngIf="item.primary" nzColor="blue">
+                  <nz-icon nzType="crown" nzTheme="outline" style="margin-right: 2px"></nz-icon>
+                  {{ 'PrimaryUser' | translate }}
+                </nz-tag>
+                <nz-tag *ngIf="item.writable" nzColor="green">
+                  <nz-icon nzType="edit" nzTheme="outline" style="margin-right: 2px"></nz-icon>
+                  {{ 'Writable' | translate }}
+                </nz-tag>
+                <nz-tag *ngIf="!item.writable" nzColor="orange">
+                  <nz-icon nzType="eye" nzTheme="outline" style="margin-right: 2px"></nz-icon>
+                  {{ 'ReadOnly' | translate }}
+                </nz-tag>
+              </div>
+            </div>
+
+            <div style="margin-left: 12px" [style.display]="item.primary ? 'none' : 'block'">
+              <button
+                nz-button
+                nzDanger
+                nz-popconfirm
+                nzType="text"
+                nzSize="small"
+                nzPopconfirmPlacement="topRight"
+                nzPopconfirmOverlayClassName="share-user-delete-button"
+                [disabled]="item.primary"
+                [nzPopconfirmTitle]="'RemoveShareUserConfirm' | translate"
+                (nzOnConfirm)="onDeleteShareUser(item)"
+                style="border: none; box-shadow: none"
+              >
+                <nz-icon nzType="close" nzTheme="outline"></nz-icon>
+              </button>
+            </div>
+          </div>
+        </nz-card>
+      }
+    </div>
+  </div>
+
+  <button
+    nz-button
+    nzBlock
+    nzType="primary"
+    class="share-button"
+    (click)="onShowCreateShareLinkModal()"
+    style="border-radius: 4px; height: 35px"
+  >
+    <nz-icon nzType="usergroup-add" nzTheme="outline"></nz-icon>
+    {{ 'CreateLink' | translate }}
+  </button>
+</section>
+
+<nz-modal
+  nzCentered
+  [(nzVisible)]="showCreateShareLinkModal"
+  [nzTitle]="modalTitle"
+  [nzContent]="modalContent"
+  [nzFooter]="modalFooter"
+  (nzOnCancel)="handleCancel()"
+  nzWrapClassName="create-share-link-modal"
+>
+  <ng-template #modalTitle>
+    @if (showCreateShareLinkForm) {
+      {{ 'CreateLink' | translate }}
+    } @else {
+      {{ 'ShareLink' | translate }}
+    }
+  </ng-template>
+
+  <ng-template #modalContent>
+    @if (showCreateShareLinkForm) {
+      <form nz-form nzLayout="vertical" [formGroup]="shareLinkForm">
+        <nz-form-item>
+          <nz-form-label>{{ 'ExpiredTime' | translate }}</nz-form-label>
+          <nz-form-control>
+            <nz-select formControlName="expired_time">
+              @for (option of shareExpiredOptions; track option.value) {
+                <nz-option [nzValue]="option.value" [nzLabel]="option.label"></nz-option>
+              }
+            </nz-select>
+          </nz-form-control>
+        </nz-form-item>
+
+        <nz-form-item>
+          <nz-form-label>{{ 'ActionPerm' | translate }}</nz-form-label>
+          <nz-form-control>
+            <nz-select formControlName="action_permission">
+              @for (permission of shareUserPermissions; track permission.value) {
+                <nz-option [nzValue]="permission.value" [nzLabel]="permission.label"></nz-option>
+              }
+            </nz-select>
+          </nz-form-control>
+        </nz-form-item>
+
+        <nz-form-item>
+          <nz-form-label>{{ 'ShareUser' | translate }}</nz-form-label>
+          <nz-form-control>
+            <nz-select
+              nzShowSearch
+              nzAllowClear
+              nzServerSearch
+              nzMode="multiple"
+              formControlName="users"
+              (nzOnSearch)="debounceSearch($event)"
+            >
+              @if (!loading) {
+                @for (o of searchedUserList; track o) {
+                  <nz-option [nzValue]="o" [nzLabel]="o.username"></nz-option>
+                }
+              } @else {
+                <nz-option nzDisabled nzCustomContent>
+                  <nz-icon nzType="loading" class="loading-icon" />
+                  Loading Data...
+                </nz-option>
+              }
+            </nz-select>
+          </nz-form-control>
+        </nz-form-item>
+      </form>
+    }
+
+    @if (showLinkResult) {
+      <form nz-form nzLayout="vertical">
+        <nz-form-item>
+          <nz-form-label>{{ 'ShareLink' | translate }}</nz-form-label>
+          <nz-form-control>
+            <input nz-input [value]="shareLink" />
+          </nz-form-control>
+        </nz-form-item>
+
+        <nz-form-item>
+          <nz-form-label>{{ 'ShareCode' | translate }}</nz-form-label>
+          <nz-form-control>
+            <input nz-input [value]="shareCode" />
+          </nz-form-control>
+        </nz-form-item>
+      </form>
+    }
+  </ng-template>
+
+  <ng-template #modalFooter>
+    <button nz-button nzType="default" (click)="handleCancel()">{{ 'Cancel' | translate }}</button>
+
+    @if (showCreateShareLinkForm) {
+      <button nz-button nzType="primary" (click)="onCreateShareLink()">
+        {{ 'ConfirmCreation' | translate }}
+      </button>
+    }
+
+    @if (showLinkResult) {
+      <button nz-button nzType="primary" (click)="copyLink()">{{ 'Copy' | translate }}</button>
+    }
+  </ng-template>
+</nz-modal>
