+++ src/app/elements/drawer/session-share/session-share.component.scss
@@ -1,0 +1,65 @@
+.share-user-list {
+  height: calc(100vh - 195px);
+
+  .ant-tag {
+    margin: 0;
+    font-size: 11px;
+    padding: 0 6px;
+    height: 20px;
+    line-height: 18px;
+    border-radius: 4px;
+  }
+
+  .ant-divider {
+    margin: 0 0 12px 0;
+    border-top-color: var(--el-divider-border-color);
+
+    ::ng-deep .ant-divider-inner-text {
+      display: flex;
+      align-items: center;
+      gap: 8px;
+      padding: 0 12px;
+      font-size: 14px;
+      font-weight: 500;
+      color: var(--el-text-color-light);
+    }
+
+    .ant-list {
+      .ant-list-item-meta {
+        align-items: center;
+
+        .ant-list-item-meta-content {
+          .ant-list-item-meta-description,
+          .ant-list-item-meta-title {
+            margin-bottom: unset;
+            color: var(--el-text-color-light);
+          }
+        }
+      }
+
+      .ant-list-item-action li {
+        color: var(--el-text-color-light);
+
+        svg {
+          transition: color 0.3s ease;
+
+          &:hover {
+            cursor: pointer;
+            color: red;
+          }
+        }
+      }
+    }
+  }
+}
+
+::ng-deep .create-share-link-modal {
+  .ant-modal-content {
+    border-radius: 4px;
+    overflow: hidden;
+
+    .ant-modal-header {
+      border-radius: 8px 8px 0 0;
+    }
+  }
+}
